+++ packages/core/src/StateMachine.ts
@@ -49,6 +49,7 @@ import type {
 } from './types.ts';
 import { isErrorActorEvent, resolveReferencedActor } from './utils.ts';
 import { createActor } from './interpreter.ts';
+import isDevelopment from '#is-development';
 
 export const STATE_IDENTIFIER = '#';
 export const WILDCARD = '*';
@@ -182,7 +183,12 @@ export class StateMachine<
     };
     this.version = this.config.version;
     this.types = this.config.types ?? ({} as any as this['types']);
+
     this.transition = this.transition.bind(this);
+    this.getInitialState = this.getInitialState.bind(this);
+    this.restoreState = this.restoreState.bind(this);
+    this.start = this.start.bind(this);
+    this.getPersistedState = this.getPersistedState.bind(this);
 
     this.root = new StateNode(config, {
       _key: this.id,
@@ -193,6 +199,18 @@ export class StateMachine<
 
     this.states = this.root.states; // TODO: remove!
     this.events = this.root.events;
+
+    if (
+      isDevelopment &&
+      !this.root.output &&
+      Object.values(this.states).some(
+        (state) => state.type === 'final' && !!state.output
+      )
+    ) {
+      console.warn(
+        'Missing `machine.output` declaration (top-level final state with output detected)'
+      );
+    }
   }
 
   /**
@@ -596,6 +614,7 @@ export class StateMachine<
 
       const actorRef = createActor(logic, {
         id: actorId,
+        parent: _actorCtx?.self,
         state: actorState
       });