+++ packages/xstate-solid/test/useMachine.test.tsx
@@ -78,10 +78,10 @@ describe('useMachine hook', () => {
           initial: 'done',
           states: {
             done: {
-              type: 'final',
-              output: 'persisted data'
+              type: 'final'
             }
-          }
+          },
+          output: 'persisted data'
         }) as any
       }
     })